(Rawhide/F28) Drop --noboot from container image creation
When building Fedora Server base images (such as when building F27
Modular Server), the --noboot option results in the container image
attempting to mount /boot with XFS like the rest of the system.
This results in the image-creation failing.

Since the partitions don't matter in the end (the files are tarred
up and shipped that way), we'll skip this optimization.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>