Don't exclude libusbx from container images, libdnf needs it
This exclusion has never actually worked. Look at a successful
F27 container-minimal build:

https://koji.fedoraproject.org/koji/taskinfo?taskID=25064051

If you check one of the tasks and look at the oz log, it shows
that libusbx is actually installed.

This is because both dnf and microdnf require libdnf, which
requires librepo, which requires gpgme, which requires gnupg2,
which requires libusb.

In Fedora 27, anaconda/dnf handle this by ignoring the attempt
to exclude libusbx and just installing it anyway.

In Rawhide, however, anaconda/dnf behaviour is different. I
don't know when it changed, but now anaconda/dnf honor the
kickstart and exclude libusbx from the install transaction...
which means the image build just fails, because the deps for
dnf/microdnf cannot be satisfied. So we should just ditch the
exclusion, it's bogus. See a failed Rawhide build attempt:

https://koji.fedoraproject.org/koji/taskinfo?taskID=25077542

Signed-off-by: Adam Williamson <awilliam@redhat.com>