workstation: Don't autostart gnome-software on live media
gnome-software session service automatically downloads yum metadata,
which all goes to the RAM-backed filesystem overlay when running the
live image. This is undesired as it can make it difficult to install
Workstation on low memory devices.

To fix this, this commit disables gnome-software xdg autostart service
and the gnome-shell search provider on the live media.

https://bugzilla.redhat.com/show_bug.cgi?id=1560504