Add initscripts to fedora-live-base.ks
It was removed from the @core group in fedora-comps d7faeb5,
but we are still using stone-age sysv-style scripts for the
early boot live image setup stuff (livesys and livesys-late),
so we need it installed in live images. Without this, lives are
utterly broken, because livesys fails almost immediately and
does not do any of the expected live customizations.

Signed-off-by: Adam Williamson <awilliam@redhat.com>