Add chkconfig package to fedora-arm-base.ks
Both the kickstart itself and imgcreate expect chkconfig to be
available in the installed system, so we'd better make sure it
is. Up till Fedora 30 Beta it was getting pulled in somehow;
since Fedora-30-20190330.n.3 it seems it is not (probably not
in Rawhide either). In any case, it's clearly correct to list it
explicitly.

See https://bugzilla.redhat.com/show_bug.cgi?id=1695637

Signed-off-by: Adam Williamson <awilliam@redhat.com>