Update Workstation favorite apps for live session
We haven't had Evolution or Shotwell installed by default in a long time, so this list is pretty stale. Sync it with upstream, except for Geary (which we don't ship) and Music (we still have Rhythmbox). I'll also omit GNOME Software, since I'm not sure we can expect installing apps to work well in a live session.