fedora-live-base: Exclude fcoe-utils and device-mapper-multipath from %packages
As discussed here:
https://fedoraproject.org/wiki/Changes/RemoveDeviceMapperMultipathFromWorkstationLiveCD
And on the anaconda devel list, we do not want device-mapper-multipath
on the livecd / on installs done from the livecd, because rather then
properly supporting hotplug udev add/change events it brings in
the long obsolete systemd-udev-settle.service which causes significant
delays during boot.

Since multipath is very much a data-center / server only feature we
can simply exclude it from the livecd to get rid of the dependency on
systemd-udev-settle.service .

We also need to exclude fcoe-utils since that depends on
device-mapper-multipath. Since FCOE too is a data-center / server
only feature, we can simply exclude fcoe-utils too.