Do not use fedora-container-common.ks in fedora-eln-container-base.ks
The Fedora ELN container is special in a way that the "%{fedora}"
RPM macro is not defined there. This breaks the Fedora ELN container
build, because fedora-container-common.ks expects this macro to exist.

This commit merges fedora-container-common.ks into
fedora-eln-container-base.ks and hardcodes Fedora 34 GPG key there.

The fedora-eln-container-base.ks will be updated once we move to
newer Fedora.

Signed-off-by: Jan Kaluza <jkaluza@redhat.com>